progettazione_2/progettino/requirements.txt
Flask>=2.0
Psycopg2-binary ==2.9.9
Werkzeug>=2.0